Init lib package with factorize module
The lib package holds utility and math functions, used in multiple
solutions with embedded unit tests. This and the code templates in
template.py and runner.py are now documented in the README.
Add a .gitignore entry for the hypothesis cache folder. Hypothesis is used
by unittests in the lib package.